+++ iOS/Village/Village/Presentation/ChatRoom/ChatRoomViewController.swift
@@ -9,50 +9,6 @@ import UIKit
 
 class ChatRoomViewController: UIViewController {
     
-    let postView: UIView = {
-        let view = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        
-        return view
-    }()
-    
-    let postImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.image = UIImage(systemName: ImageSystemName.photo.rawValue)
-        imageView.contentMode = .scaleAspectFit
-        imageView.backgroundColor = .primary100
-        imageView.layer.cornerRadius = 16
-        
-        return imageView
-    }()
-    
-    let postTitleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.text = "닌텐도 스위치"
-        label.font = .systemFont(ofSize: 16, weight: .bold)
-        
-        return label
-    }()
-    
-    let postPriceLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.text = "10000원"
-        label.font = .systemFont(ofSize: 14, weight: .bold)
-        
-        return label
-    }()
-    
-    let postAccessoryView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.image = UIImage(systemName: ImageSystemName.chevronRight.rawValue)
-        
-        return imageView
-    }()
-    
     let chatStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
@@ -86,6 +42,8 @@ class ChatRoomViewController: UIViewController {
         
         return button
     }()
+    
+    let postSummaryView = PostSummaryView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -97,12 +55,7 @@ class ChatRoomViewController: UIViewController {
     }
     
     private func setUI() {
-        
-        postView.addSubview(postImageView)
-        postView.addSubview(postTitleLabel)
-        postView.addSubview(postPriceLabel)
-        postView.addSubview(postAccessoryView)
-        view.addSubview(postView)
+        view.addSubview(postSummaryView)
         
         chatTextField.delegate = self
         
@@ -115,35 +68,6 @@ class ChatRoomViewController: UIViewController {
     }
     
     private func configureConstraints() {
-        
-        NSLayoutConstraint.activate([
-            postView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            postView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            postView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            postView.heightAnchor.constraint(equalToConstant: 100)
-        ])
-        
-        NSLayoutConstraint.activate([
-            postImageView.topAnchor.constraint(equalTo: postView.topAnchor, constant: 10),
-            postImageView.leadingAnchor.constraint(equalTo: postView.leadingAnchor, constant: 10),
-            postImageView.widthAnchor.constraint(equalToConstant: 80),
-            postImageView.heightAnchor.constraint(equalToConstant: 80)
-        ])
-        
-        NSLayoutConstraint.activate([
-            postTitleLabel.topAnchor.constraint(equalTo: postView.topAnchor, constant: 25),
-            postTitleLabel.leadingAnchor.constraint(equalTo: postImageView.trailingAnchor, constant: 20)
-        ])
-        
-        NSLayoutConstraint.activate([
-            postPriceLabel.topAnchor.constraint(equalTo: postTitleLabel.bottomAnchor, constant: 10),
-            postPriceLabel.leadingAnchor.constraint(equalTo: postImageView.trailingAnchor, constant: 20)
-        ])
-        
-        NSLayoutConstraint.activate([
-            postAccessoryView.topAnchor.constraint(equalTo: postView.topAnchor, constant: 39),
-            postAccessoryView.trailingAnchor.constraint(equalTo: postView.trailingAnchor, constant: -16)
-        ])
         
         NSLayoutConstraint.activate([
             chatStackView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
